+++ springboot-trading-data/src/main/java/de/segoy/springboottradingdata/repository/ConnectionDataRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ConnectionDataRepository extends CrudRepository<ConnectionData, Integer> {
+public interface ConnectionDataRepository extends CrudRepository<ConnectionData, Long> {
 
     @Modifying
     @Query("update ConnectionData c set c.connected = false where c.id = ?1")
-    void setConnectFalseById(Integer id);
+    void setConnectFalseById(Long id);
 }
